refactor: moved buttons to separate component, more prop grouping, cleaned up code for readability
