+++ frontend-next-migration/src/preparedPages/ClanPages/ui/_components/clanoverview/buttonField/index.ts
@@ -1,0 +1,1 @@
+export { default as ButtonField } from "./ButtonField";
